+++ gamma_dashboard/static/dashboard/js/app/lib/ConfigPreset.js
@@ -1,0 +1,22 @@
+const path = require('path');
+const resolveFilepaths = require('./resolveFilepaths');
+
+const defaultConfigDir = path.resolve(__dirname, '../config');
+
+function ConfigPreset({
+  defaultDir = defaultConfigDir,
+  defaultFilename,
+  searchFilenames,
+  searchFilepaths,
+}) {
+  return {
+    get resolvedFilepath() {
+      return resolveFilepaths(
+        searchFilenames.map(filename => `./${filename}`),
+        [...searchFilepaths, defaultDir],
+      );
+    },
+  };
+}
+
+module.exports = ConfigPreset;