+++ gamma_dashboard/static/dashboard/js/app/src/components/LogoDropdown/components/index.js
@@ -1,0 +1,2 @@
+export { default as FeedbackForm } from './FeedbackForm';
+export { default as Modal } from './Modal';