+++ gamma_dashboard/static/dashboard/js/app/webpack.config.js
@@ -26,11 +26,7 @@ module.exports = {
         },
         {
           test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/, /\.svg$/],
-          loader: require.resolve('url-loader'),
-          options: {
-            limit: parseInt('10000'),
-            name: 'images/[name].[ext]',
-          },
+          loader: 'url-loader',
         },
       ]
     },